[YS-258] refactor: 공고목록 컴포넌트로 분리하여 컴포넌트 복잡도 개선
