+++ src/app/home/components/PostContainer/PostContainer.css.ts
@@ -15,24 +15,6 @@ export const postContainerTitle = style({
   color: colors.text06,
 });
 
-export const postCardContentContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '1.6rem',
-});
-
-export const postCardContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '0.6rem',
-  minHeight: '40rem',
-});
-
-export const totalPostCount = style({
-  ...fonts.label.large.R14,
-  color: colors.text03,
-});
-
 export const filterWrapper = style({
   display: 'flex',
   alignItems: 'center',
@@ -42,12 +24,4 @@ export const filterWrapper = style({
 export const recruitCheckLabel = style({
   ...fonts.label.large.SB14,
   color: colors.text06,
-});
-
-export const watchMoreButton = style({
-  ...fonts.label.large.SB14,
-  color: colors.text04,
-  backgroundColor: colors.field04,
-  padding: '1.2rem 0',
-  borderRadius: '1.2rem',
 });
